requirements.txt
# Core dependencies
openai>=1.12.0
PyQt6>=6.6.0
PyQt6-Qt6>=6.6.0

# Audio processing
sounddevice>=0.4.6
numpy>=1.24.0

# File processing
python-docx>=1.1.0

# Configuration and utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Logging
colorlog>=6.8.0

# Async support
aiohttp>=3.9.0
asyncio

